pyproject.toml
[project]
name = "mcp-bridgekit"
version = "0.6.0"
description = "Embeddable MCP stdio → HTTP bridge with timeout survival"
readme = "README.md"
license = {text = "MIT"}
requires-python = ">=3.11"
dependencies = [
    "fastapi>=0.115.0",
    "uvicorn[standard]>=0.34.0",
    "mcp>=1.0.0",
    "redis>=5.2.0",
    "rq>=2.0.0",
    "pydantic-settings>=2.7.0",
    "structlog>=24.4.0",
    "httpx>=0.28.0",
    "jinja2>=3.1.0",
]

[project.optional-dependencies]
dev = ["pytest>=8.0.0", "ruff", "black"]

[project.scripts]
mcp-bridgekit-worker = "mcp_bridgekit.worker:main"

[build-system]
requires = ["setuptools>=45", "wheel", "setuptools-scm"]
build-backend = "setuptools.build_meta"

[tool.setuptools.packages.find]
where = ["src"]
